Move most of lib.rs to a new file raw_cactus.rs.
